Change waterskin add/remove behaviour.
If the script adds a waterskin of X litres, we will overwrite the waterskin that's already there, if any (assuming the requested waterskin and the one the player has are both small or both large).

If the script tries to remove a waterskin of X litres, the player will only lose their waterskin if it has EXACTLY X litres. If their waterskin has Y litres, it will remain in the inventory.