Big audio & audio settings cleanup.
Remove "Sound Enabled" setting, as it would disable BASS if used, leading to a load of warning messages and missing sounds when re-enabled.

Currently-playing track(s) now get quieter/louder when using the music volume slider in the audio settings menu.

Encapsulate GlobalMusicVolume and GlobalFXVolume in functions (at the time of writing this is only strictly necessary for music volume, since SetVolumeMusic needs to also set the volume of currently-playing tracks).

Make "apply" button return to the previous menu, as it does with keyboard controls.

Replace a printf call with a TENLog call.