Update tr3_tony.cpp
Currently , on frame 120 of Tony's dying animation there is a call for a hardcoded sound (warp.wav / 76: Blast Circle) to be played. I cannot fix this in wadtool and it is causing sound issues such as tony's dying sound to be played when lara fires her shotgun. SO I've commented out this line to see if that fixes it.