Remove concept of temporary Moveables - it won't play sensibly with saves if the lifetime of an ITEM_INFO is linked to the Lua variable.
Make SetName return a bool so we know whether or not it succeeded.

Add Destroy(), which allows a user to kill a moveable from the script.

Add calls to RemoveMoveableFromMap and AddMoveableToMap in places where a Moveable needs to be linked or delinked from an ITEM_INFO.

Stop initialising a Moveable if an object with that name already exists in the name map.