Moving the Range attacks as constant expressions
- Dog: Only moving the range attack as constexprs.

- Small Scorpion: Moving the range attack, and the attack damages as constexprs.

- Big Scorpion: Moving the range attack and the radius as constexprs.

As a big creature, looks like it uses an offset to help its navigation on corners. I think it could be considered the radius of the creature.

Doubt:

In line 108, there is a number: 1344, I think it's an angle and somehow is related to the roll rotation of the scorpion during the ramps, but I'm not very sure of how it work so I prefered don't touch that one.