Update Attack Ranges Names and Big Scorpion.
Dog: as was suggested, I've changed the name to be more alike the other values of entities.

SmallScorpion: as was suggested, I've changed the name to be more alike the other values of entities.

Big Scorpion: as was suggested, I've changed the name to be more alike the other values of entities.

The orientation issue:
I'll try to be brief here, the thing is that after styding Big_Scorpion code, I see the Scorpion orientation calculation method is conflicting with the orientation calculation method from "CreatureAnimation" function.

I think is likely that all these lines calculating tilts and roll in the Big Scorpion will get deleted on a future update for being a deprecated mehod.

So I changed the BIG_SCORPION_RADIUS back to the simple number in the lines, (as it'snot used for anything else) and the 1344 I put it as SECTOR (1.34) for a style suggestion, although again, is very likely this line get replaced by the new method.