Independent Guide update
Changes:

* New use of the variables:
ItemFlags[2]
ItemFlags[4]

* GoalNode
Changed all Lara.Location by a local variable GoalNode, this variable gets the value of Lara.Location (all allies uses the same) or ItemFlags[4] (is personal for this guide). The assigment depends of the value in ItemFlags[2]. (If 0, (default) it will keep using Lara.Location to keep it working like legacy TR4, if 1 or bigger, will use the personal ItemFlags[4]

* Ignore waiting
In the actions, read inscription, crouching, and activate trap, the condition has been extended, so if ItemFlags is 0 or 1, the guide will keep waiting for Lara to be closer than 2 sectos before do the action, but if ItemFlags[2] is 2 or higher, the guide will do the action as he arrives.