Guide, cleaning, polishment, fixes and implementation of new feature.
Some cleaning, polishment and bug fixes (It wasn't refresh the AI searched when moving to a passed node).

Also implemented a new feature, the run default, if the flag is active (with the Lua function) the guide will go running to the next node (unless the next node were very close).