Scary text flag
Now, when the flag is active, the Guide will get scared when he reads the next inscription.

Like in TR4, but without need to be in room 70.

The flag onces the action is done, it deactivates itself.

Unlike TR4, I set that the scared face changed back to his normal one at the end of the reading animation, so builders can keep using it with his normal face.