fix_falling_block
If the falling block was placed over a portal, it was breaking into pieces the moment Lara stepped on it.

I was trying to debug the function GetDistanceToFloor but my game had crashes when I tried to use it. So I used a different function to get the floor Distance.

In this way I could get a reliable floor position to calculate a distance, and the falling blocks don't break anymore on the instant that Lara jumps on them, but after a delay, as it was designed.