Update tr5_twoblockplatform.cpp
Problem: When the platform reaches to a vertical portal, Lara and enemies that were over it, fall through the platform back to the floor.

The problem looked to be related with the bridges system, and the creatures anti-digging mechanic (the one that push the enemy up or down to ensure it doesn't get digged in the floor or ceiling).

After several approaches, this one is the less bad that I've found. It just ensures the platform bridge gets updated in the frames before to reach the portal.

Doing that, makes the bug stop happening. Lara and enemies can stay over the platform without disruption, falls nor trembles.

However UpdateBridgeItem happens to be an expensive function, and this approach makes the function trigger during 1-15  times depending of the platform speed plus the other 2 times when reaches the portal. (to delete the bridge in the old room and create other in the new).

I could not find a better way, but I could find a way.