Update falling_block.cpp
Changed fix approach.

I could not fix the GetDistanceToFloor when platform is on the lower portal height.

So instead, I moved the segment that managed the collision of the block with the floor, and I put it on Fall routine.

Then the problem also dissapears because for the crumbling stage doesn't get interrupted because it does not check for the floor anymore, only when it start to fall.