Update cog_switch.cpp
Lara Ungrab Animation better fix.

This removes the forced animation call and leave the system do the animation automatically setting the new State when the action key is released.

It had to be removed from the lines 159_162 because these lines only are activated when the wheel finishes its animation (What unable Lara to do the animation change on time, making her do the PULL animation one more time).