Update cog_switch.cpp
Testing this, I found an issue.

It was happening that when player releases action key, sometimes Lara repeate the PULLING animation one last time. But the cog was remaining static.

This commitment fixes that.

The issue was happening when the player drop the key after the frame animation check (between PULL frames 36- 54) Then, The state change do not happen, so Lara does the PULL animation one last time.

But previously when player release the key and the wheel reach to its last frame (53), this stopped.

These changes add an extra condition to the wheel, check if in the last frame, Lara is performing the RELEASE animation, then it stops.

Otherwise, the wheel is turned On again, to perform one more round Like Lara is doing.