Bug fix on FixedCameras.
Bug description: When a fixed camera was being triggered, and, if the target go behind and obstacle, then the screen turned black.

Reason: In those conditions, the move camera check the origin vector received, however, that vector was incomplete because it had the room 0, so although Move camera corrected the room later, it was already using some wrong calculus at room 0 that caused the black camera (this got an offset that places it out of the real room).

The reason why the vector was getting room 0, was because the operator function = of the GameVector class, was forgetting to copy the Room Number.

After this fix, the bug is not observed anymore.