Lua new functions adngel (#934)
* Function: Update Floor Collision

Adds a new function to allow update the Floor collision of some objects (bridges, trapdoors, platforms, etc).

* Function: Play Camera

Allows to set activate a camera while this function is called.

Right now, it requires the position and room of the target. (WIP)

* Update

* Removed the duplicated comment for UpdateFloorCollision.
* Added new function GetIndex to the class Moveable.
* Added new function GetIndex to the class CameraObject.
* Added a new variable to the object LevelCameraInfo (Index).
* Edited the function PlayCamera

* Function: GetCameraType

* Update descriptions

Updated descriptions for.
* GetItemFlags(index)
* SetItemFlags (value, index)

* Bug fix on FixedCameras.

Bug description: When a fixed camera was being triggered, and, if the target go behind and obstacle, then the screen turned black.

Reason: In those conditions, the move camera check the origin vector received, however, that vector was incomplete because it had the room 0, so although Move camera corrected the room later, it was already using some wrong calculus at room 0 that caused the black camera (this got an offset that places it out of the real room).

The reason why the vector was getting room 0, was because the operator function = of the GameVector class, was forgetting to copy the Room Number.

After this fix, the bug is not observed anymore.

* Removal of Function UpdateFloorCollision

* Erasing of forgotten line

It was included for the UpdateBridgeFunction, but as that function is not here anymore, this include is not necessary.

* GetIndex functions hidden from public lua API

They still remain as they are being used by other classes, but they can't be called by lua script anymore.

* First attempht of new Enum for lua

Need to test first.

* Removal of GetIndex() in CameraObject

The GetIndex from the moveable class, is used in CameraObject class.

But this one from CameraObject, although I created it was still unused. So I'm unmaking it.

The Index var in the LevelCameraInfo.cpp is still there and is required for PlayCamera() function to work.

* Update GetCameraType() @usage example.

* Branch Update

- Extracted CameraType enumerator to its own header.
- Fixed formatting of CameraObject.cpp line 70

* Added comments for the doc

- Added comments to the enumerator for the doc.
- Removed datar from the function description, guiding to the enumerator CameraType info.

* Update Changes.txt

Added the content of this branch to the changes.txt

* Error correction in CameraTypes enumerator header.

* Updated usage example.

Updated for a shorter example.

* Update Changes.txt

Co-authored-by: Lwmte <3331699+Lwmte@users.noreply.github.com>