Fix Big Rat death crash
It was happening, that when the Big Rat was killed, like other objects, "CreatureAnimation" puts it in deactivate state and deletes its item->Data.

But Control function was trying to access that item->Data later in the "IsRatOnWater" function, that must be activated after CreatureAnimation. As item->Data didn't exist, the game crashes.

This fix what it does is add a check verification to IsRatOnWater so if it detects that item->Data is invalid, returns the value.

(For the return, I decided move some lines to still provide a similar comparison of waterlevel != of NoHeight to avoid other issues).