Added flipeffect 47
Flipeffect 47 ActiveNextAnimOnLaraSlot

Can be used with LARA_EXTRA_ANIMs so when the animation ends, if this function is call, (by Animcommand 47), it can use the value in Next Animation to activate it in LARA.

So if Lara_Extra_Anim next animation is 11.
Without this function, the engine will activate the animation 11 of Lara_Extra_Anim. (Gong strike), with this function it can activate the animation 11 of LARA (idle state).