Updated Camera:SetPosition
Created a function to for the fixed camera position Update for those cases where this change (by lua) but there are not other game events to refresh the change.

For example: If Lara is moving, the camera is constantly refreshing, she can step triggers that change camera and this will work fine.

But if Lara is stopped, no keyboard input comming, and the fixed camera change its position externally (lua, evensequence, etc) these wouldn't be refreshed in the camera until a game Event happens (like player pressing any key).

This commitmment aims to solve that issue.