Todo note
When the civvy is in function, he searches for active entities nearby and if conditions accepts, he selects it as target.

However, if an AI object is applied, that enemy search is not done, to avoid it mess with the AI objects searches that are also stored in the same creature->enemy (variable, where he also stores his target).

Because of this, when is in a AI state (PATROL, GUARD), there is not enemy search detection, so he won't snap out of the AI state unless that he get actually attacked (a condition that make him alert, and being alert, the AI states dissapear).

I leave pending for the future the polishment of this feature for this creature (that surelly will must be done in addition with other creatures).

For the others enemies, this problem is also there but is more subtle because they only have to check for Lara, but this enemy that use to be friendly to Lara, need to focus on the other entities, and the function GetAITarget is not ready for that at this moment.