Fix bug with collision detection
When height ocb is set to make the pushable walkable. This creates a bridge collision and the pushable starts to use that collision system instead of the normal object one.

However in the test rutine, specifically during the moment when it checks if there are space for Lara. It was doing a collision detection that was being affected by the bridge collision of the pushable. That was causing false positives that in the practice was unabling Lara to pull from the objects in several circumstances.