Update ElectricCleaner.cpp
I can't use the bounding box for the height, because due to the wires, the box height varies, and gets also bigger than the machine, so I changed it back to a fixed BLOCK(1) size.

Height variable is not needed anymore as it won't be changed, so I passed the constant value directly.