Deleted ClearMovableBlockSplitters function
It's purpose was unclear, the way it used wasn't optimal nor ordered, and it showed a risk of potential infinite recursivity if there were several blockabe sectors at the same height that were not connected. (Unusual case, but the time bomb is there).

So at the end I decided to discard it. If it really were needed to block the blockages tiles, it can be done with the Stopper flag which does the same with less fusses.