string:SetTranslated() method and isStringDisplaying() function.
The string:SetTranslated() method sets the translated parameter. If it is true, this will be the key string for the translation that will be displayed. If it is false, this will be the string that will be displayed.
The isStringDisplaying() function checks whether the string is displayed. It returns the bool true if it is shown, false if it is hidden.