String:SetTranslated() method and isStringDisplaying() function. (#1079)
* string:SetTranslated() method and isStringDisplaying() function.

The string:SetTranslated() method sets the translated parameter. If it is true, this will be the key string for the translation that will be displayed. If it is false, this will be the string that will be displayed.
The isStringDisplaying() function checks whether the string is displayed. It returns the bool true if it is shown, false if it is hidden.

* DisplayString::SetTranslated @tparam

Corrected @tparam argument name of DisplayString::SetTranslated() method.

* Corrected method name.

capitalized the first letter of the method name
StringsHandler::IsStringDisplaying()

* Corrected method name.

capitalized the first letter of the string:IsStringDisplaying() method name.

* Corrected method name.

Capitalized the first letter of the method name
StringsHandler::IsStringDisplaying()

* Update DisplayString.cpp

code improvements

* code improvements

code improvements

* code improvements

code improvements