Bug Fix
The problem of the displaced text was caused because the new pos point was being calculated after the pointColl.

pointColl was not just to get the height, is the actual point that will check for the flags, so updating the pos X-Z after the GetCollision(), was causing it to check an old frame (That's why of the displacement).

And when Lara move away of the range border, that move the pos out of bounds what cause the big offset translation of the bug.

After this change, I didn't appreciate that error anymore.

Also because now it checks the right sector, the previous commit fix was not needed anymore so this unmade that too.