Fix Pushable Alignment issue
Looks like that control updates interfere with the alignment process. So I paused the update if the alignment is happening.

Also changed the polarity of AlignCenter. Now if nothing is added, it will align in the center like TR4, only when the +4 OCB is added, Lara will can grab the pushable freely.