Fix Snowmobile baddie crash
In this item, Data is being used as creatureInfo so it can't store the short value of the snowmobile item index.

Therefore I changed the store variable to the ItemFlags[0] so the game won't crash anymore when it triggers this enemy. (However, it still doesn't work as he rides towards the South West direction of the map ignoring Lara location).