Added DisplayStringWrapper
This commit tries to solve the issue of can call two different functions (CreateString and CreateStringDeprecated) using the same lua instruction (DisplayString).

The problem was that both functions use different arguments, so the former implementation that used sol::overload was causing errors in this situation.

The implementation provided in this commit, is use one function DisplayStringWrapper linked to the lua DisplayString. That is open to a variety of arguments (reading them as objects instead of types). So the user in lua can provide any of the two version (the new one or the deprecated one).

And later this wrapper, will read the arguments, and will sort them to call the new method or, if arguments comes in the deprecated way, will sort them to make them work with the new function.

(Therefore, this commitment also deletes the now unused CreateStringDeprecated function)