Reformat: From header to static methods
Changed the auxiliar functions to static methods defined uniquely in the source code. And deleted from the header code.

As these functions are unique for the Electric Cleaner and don't need to be known by external codes. So are moved away from the header.

For this, the order of the functions in the source code also has been changed, so now Control and Collide functions are at the bottom.