Update Slopes alignments
GetShortestAngle is converting the diferenceof slopeAngle and dirAngle to short type, (What will not take the value 32768 anymore)

Therefore the next if have been modified to check only the forward (angle 0) and backwards (angle 180).

Note: ANGLE (180) and ANGLE (-180) gives both the same result, -32768, Although ANGLE 180 is positive, the short type rounds it back to the negative side. However in the code I think will be easier to read using the positive 180 angle.