Fix: Swamp camera shaking
I explored the code a bit, but fixing one thing ended breaking another, and all that with a camera refactor in the roadmap.

So I the end I decided keep it simple with a hack to avoid the solve the current. Once the code keep improving, surelly this will can be handled better.