Update lara_fire.cpp
That break was making that the flare routines don't activate during the exception animations.

(In this animations, the flare pose in the left hand shall not activate, that's why the ControlLeft is set to false, but other effects like Light and Sparks are still needed, so it can't exit the cycle with break here yet.