Fix Swamp Rooms camera
The previous hack, worked in the quick sand room I had. But then I tested different level with differnt swamp pool and a different error arised. That took me to keep researching.

Also when I fixed a bug in the pushables related to the swamps, I learned also a couple of more things.

So this version return to the original line, but changes the g_Level.Rooms[probe.RoomNumber].y for g_Level.Rooms[probe.RoomNumber].maxceiling

In that way, it ensures the camera goes over the swamp room, independently of the pool depth.

After this I didn't got shacking camera issue anymore on any of the two pools, and it neither sink the camera into the swamp pool (as happened before with the HACK in the deep swamp pools).