Meshbit updates
So for DRAGON_FRONT, BONES_FRONT and BONES_BACK, no meshBits changes are needed as these objects have to render all its meshes.

But the DRAGON_BACK has its meshes and some extra meshes of bone spine. the bone spine (mesh 21, 23 and 23) should not render, so the meshBits have been changed only for that object.