Adngel varied fixes (#1256)
* Lara 6 click landing

Changed the freefall threshold value from 131 to 126.

In that way, Lara will pass to freefall animation a bit sooner what will activate the strong landing from the 6 clicks heights.

* Fix: Swamp camera shaking

I explored the code a bit, but fixing one thing ended breaking another, and all that with a camera refactor in the roadmap.

So I the end I decided keep it simple with a hack to avoid the solve the current. Once the code keep improving, surelly this will can be handled better.

* Fix tabs

* Update lara.h

* Fix comment and math

* Fix value

* Update flame_emitters.cpp

- Fixed the small flame alpha transition when used intermittent jet flame.
- Removed unused code
- Added references variables to replace the ItemFlags names, to enhace readibility.

* LARA_FREEFALL_VELOCITY restored

Caused the strong landing when Lara is doing a jump from 4 clicks to the ground. This was not a desired behaviour, so the value has been reverted.

* Fix: Pushables waterHeight in swamp rooms.

* Update Flame Emitter 1 code

Restored the alpha shifts to recover missing behaviour.

* Fix: Grabbing and throwing Torch from crouch state.

* Lara flare pose, excluded animations list

Lara now won't do the flare pose if the animation is in the exclussion list. (Usually animations where the logic is that Lara also use the left hand, like several switches).

* Drop torch if is swimming in water.

* Update lara_fire.cpp

* Fix formatting

* Update flame_emitters.cpp

* Update burning_torch.cpp

- Changed flame sprites
- Changed GetRandomControl by Random:: functions

* Fix pushables bug when dropped to 1 click pit

* Update burning_torch.cpp

- Changed flame spawn position
- Removed randomness of initial rotation to perform better with fire sprites.

* Fix Pushable initialized underwater

- Underwater pushables won't spawn ripples if the waterheight is bigger than the actual pushable height.
- Added a missing state transtion.

* Update lara_fire.cpp

That break was making that the flare routines don't activate during the exception animations.

(In this animations, the flare pose in the left hand shall not activate, that's why the ControlLeft is set to false, but other effects like Light and Sparks are still needed, so it can't exit the cycle with break here yet.

* Fix Swamp Rooms camera

The previous hack, worked in the quick sand room I had. But then I tested different level with differnt swamp pool and a different error arised. That took me to keep researching.

Also when I fixed a bug in the pushables related to the swamps, I learned also a couple of more things.

So this version return to the original line, but changes the g_Level.Rooms[probe.RoomNumber].y for g_Level.Rooms[probe.RoomNumber].maxceiling

In that way, it ensures the camera goes over the swamp room, independently of the pool depth.

After this I didn't got shacking camera issue anymore on any of the two pools, and it neither sink the camera into the swamp pool (as happened before with the HACK in the deep swamp pools).

* Fix formatting

* Formatting

* Entity Sounds fix

- Now it will recalculate the entity room position according to the root of their transform, instead of using the item->RoomNumber that contains an offset of 2 clicks.

- Removed the camera room ENV_FLAGS requirements.

* Update animation.cpp

* Revert anim sound fix for now

---------

Co-authored-by: Sezz <sezzary@outlook.com>