fish can now follow a path (AI_FOLLOW)
You can use any number of Ai objects for a lap. If you then want to make a round for another school of fish, you simply leave out an OCB number. This way the engine knows which numbers belong to which swarm. the first AI_FOLLOW (the one that's on the same square as the fishemitter does not count for patrolling but also needs an OCB)