Bug fix: Trex double rotation
The rotation from AI.angle was correct, but looks like it was applying the same rotation to the head and the neck (was at the end was like duplicating the value).

I've edited these two lines to add only half of the angle the head and other half to the neck.