Fix: Electric Cleaner crash
When it detected a sector with no box (a tile marked with "set non-walkable sector" editor flag) the game was crashing.

Changing the detection order solved the issue.

This issue was not happening in previous version (it was originated at some point of the past month) so I'm not registering in the changes.txt