Fix Homing Dart Emiter
When triggering, it was continously spawning darts, that was because the trigger code was resetting the item->Timer variable. This variable is aimed to be used to delay the activation or provoke a timed deactivation. Using the variable adittionaly for internal behaviour like the dart spawn frequency, may lead to uncertain results like this actual bug.

Using a separate variable (ItemFlags[1]) the dart spawn frequency timer won't be influenced by the triggers code.