level.rumble sound effect parameters exposed to user
Formally: this was hardcoded in TR5 to the klaxon from the Sinking Submarine level. I have moved this to a new sound enum in the default soundmap and removed the pitch shift as this can be set in the XML by the user.  This change has also been documented on the TombEngine website.